requirements.txt
boto3>=1.34.0
fastapi
matplotlib
pandas
geopandas
psycopg2-binary>=2.9.9
pydantic
pykrige
pyproj
rasterio
rasterstats
requests>=2.31.0
seaborn
sqlalchemy
uvicorn
